c_python/requirements.txt
numpy~=2.1.3
pillow~=11.0.0
pybind11~=2.13.6
setuptools~=75.6.0